refactor metrics/status in copy.
