+++ src/query/sql/src/planner/plans/copy.rs
@@ -96,14 +96,9 @@ pub struct CopyIntoTablePlan {
     pub enable_distributed: bool,
 }
 
-fn set_and_log_status(ctx: &Arc<dyn TableContext>, status: &str) {
-    ctx.set_status_info(status);
-    info!(status);
-}
-
 impl CopyIntoTablePlan {
     pub async fn collect_files(&self, ctx: &Arc<dyn TableContext>) -> Result<Vec<StageFileInfo>> {
-        set_and_log_status(ctx, "begin to list files");
+        ctx.set_status_info("begin to list files");
         let start = Instant::now();
 
         let stage_table_info = &self.stage_table_info;
@@ -139,7 +134,7 @@ impl CopyIntoTablePlan {
 
         let num_all_files = all_source_file_infos.len();
 
-        info!("end to list files: got {} files", num_all_files);
+        ctx.set_status_info(&format!("end list files: got {} files", num_all_files));
 
         let need_copy_file_infos = if self.force {
             info!(
@@ -149,7 +144,7 @@ impl CopyIntoTablePlan {
             all_source_file_infos
         } else {
             // Status.
-            set_and_log_status(ctx, "begin to filter out copied files");
+            ctx.set_status_info("begin filtering out copied files");
             let files = ctx
                 .filter_out_copied_files(
                     &self.catalog_name,
@@ -159,8 +154,10 @@ impl CopyIntoTablePlan {
                     max_files,
                 )
                 .await?;
-
-            info!("end filtering out copied files: {}", num_all_files);
+            ctx.set_status_info(&format!(
+                "end filtering out copied files: {}",
+                num_all_files
+            ));
             files
         };
